Moved the CommandHandlerMethod in the correct package.
