+++ cqrs-core/src/main/java/com/techbank/cqrs/core/infrastructure/CommandDispatcher.java
@@ -1,7 +1,7 @@
 package com.techbank.cqrs.core.infrastructure;
 
 import com.techbank.cqrs.core.commands.BaseCommand;
-import com.techbank.cqrs.core.domain.CommandHandlerMethod;
+import com.techbank.cqrs.core.commands.CommandHandlerMethod;
 
 /**
  * Dispatches the events to the appropriate handler method.
